+++ compat/netif.ml
@@ -1,1 +1,1 @@
-include Netchannel.Frontend.Make(Netchannel.Xenstore.Make(OS.Xs))
+include Netchannel.Frontend.Make(Netchannel.Xenstore.Make(Mirage_xenstore_transport.Xs))